+++ packages/svelte/src/image/Image.svelte
@@ -1,51 +1,18 @@
-<script lang="ts" module>
-  const subscribes = new Set<
-    (event: Event, element: HTMLImageElement) => void
-  >();
-
-  export function subscribeEvent(
-    onload: (event: Event, element: HTMLImageElement) => void,
-  ): () => void {
-    subscribes.add(onload);
-
-    const unsubscribe = () => {
-      subscribes.delete(onload);
-    };
-    return unsubscribe;
-  }
-
-  export function onLoaded(img: HTMLImageElement, event: Event) {
-    subscribes.forEach((cb) => cb(event, img));
-  }
-</script>
-
 <script lang="ts">
-  import { HTMLImgAttributes } from "svelte/elements";
+  import { onLoaded, type Props } from "./image";
 
-  type Props = {
-    src: string;
-    Options?: HTMLImgAttributes;
-    Events?: Events;
-  };
-
-  // Note: パラメータの伝播がしやすい用に1変数に集約
-  export type Events = {
-    onLoad?: (event: Event, element: HTMLImageElement) => void;
-    onError?: (event: Event, element: HTMLImageElement) => void;
-  };
-
-  const { src, Options, Events }: Props = $props();
+  const { src, opts, events }: Props = $props();
 
   function onload(event: Event) {
     const img = event.currentTarget as HTMLImageElement;
-    Events?.onLoad?.(event, img);
+    events?.onLoad?.(event, img);
     onLoaded(img, event);
   }
 
   function onerror(event: Event) {
     const img = event.currentTarget as HTMLImageElement;
-    Events?.onError?.(event, img);
+    events?.onError?.(event, img);
   }
 </script>
 
-<img {src} {...Options} {onload} {onerror} />
+<img {src} {...opts} {onload} {onerror} />