+++ packages/svelte/History.md
@@ -3,4 +3,7 @@
 ## 26-04
 
 - 04-28
-  - Imageのmoduleにsubscribeを追加
+  - Image.svelte
+    - moduleにsubscribeを追加
+    - Propsなどをexport化
+    - image.tsを追加しmodule部分を分離。exportしたものを解決できるように対応